Remove duplicate entries and improve sorting
Removed duplicate URL entries, most seemed to exist due to a name change of organsiation. Have cross referenced and kept the rows based on a whois lookup of the registrant and the official names from SSC/Govt.nz A to Z list.

Resorted dataset based on both Organisation name ascending and secondary url ascending. You'll get the websites in A-Z within the organisations A-Z.